+++ src/main/java/gregtech/api/metatileentity/BaseMetaPipeEntity.java
@@ -45,7 +45,6 @@ import gregtech.api.interfaces.tileentity.IGregTechTileEntity;
 import gregtech.api.interfaces.tileentity.IPipeRenderedTileEntity;
 import gregtech.api.net.GTPacketCreateTE;
 import gregtech.api.net.GTPacketTileEntity;
-import gregtech.api.util.GTLog;
 import gregtech.api.util.GTModHandler;
 import gregtech.api.util.GTOreDictUnificator;
 import gregtech.api.util.GTUtility;
@@ -63,12 +62,9 @@ public class BaseMetaPipeEntity extends CommonBaseMetaTileEntity
 
     public byte mConnections = IConnectable.NO_CONNECTION;
     protected MetaPipeEntity mMetaTileEntity;
-    private final int[] mTimeStatistics = new int[GregTechAPI.TICKS_FOR_LAG_AVERAGING];
-    private boolean hasTimeStatisticsStarted;
     private boolean mWorkUpdate = false, mWorks = true;
-    private byte mColor = 0, oColor = 0, oStrongRedstone = 0, oRedstoneData = 63, oTextureData = 0, oUpdateData = 0,
-        mLagWarningCount = 0;
-    private int oX = 0, oY = 0, oZ = 0, mTimeStatisticsIndex = 0;
+    private byte mColor = 0, oColor = 0, oStrongRedstone = 0, oRedstoneData = 63, oTextureData = 0, oUpdateData = 0;
+    private int oX = 0, oY = 0, oZ = 0;
     protected Node node;
     protected NodePath nodePath;
 
@@ -165,132 +161,97 @@ public class BaseMetaPipeEntity extends CommonBaseMetaTileEntity
     }
 
     @Override
-    public void updateEntity() {
-        super.updateEntity();
-
+    public void updateEntityProfiled() {
         if (!hasValidMetaTileEntity()) {
             if (mMetaTileEntity == null) return;
             mMetaTileEntity.setBaseMetaTileEntity(this);
         }
 
-        long tTime;
-        if (hasTimeStatisticsStarted) {
-            tTime = System.nanoTime();
-        } else {
-            tTime = 0;
-        }
         boolean isServerSide = isServerSide();
-        try {
-            if (hasValidMetaTileEntity()) {
-                if (mTickTimer++ == 0) {
+        if (hasValidMetaTileEntity()) {
+            if (mTickTimer++ == 0) {
+                oX = xCoord;
+                oY = yCoord;
+                oZ = zCoord;
+                if (isServerSide) checkDropCover();
+                else {
+                    requestCoverDataIfNeeded();
+                }
+                worldObj.markTileEntityChunkModified(xCoord, yCoord, zCoord, this);
+                mMetaTileEntity.onFirstTick(this);
+                if (!hasValidMetaTileEntity()) return;
+            }
+
+            if (isClientSide()) {
+                if (mColor != oColor) {
+                    mMetaTileEntity.onColorChangeClient(oColor = mColor);
+                    issueTextureUpdate();
+                }
+
+                if (mNeedsUpdate) {
+                    worldObj.markBlockForUpdate(xCoord, yCoord, zCoord);
+                    mNeedsUpdate = false;
+                }
+            }
+            if (isServerSide && mTickTimer > 10) {
+                if (!doCoverThings()) return;
+
+                final byte oldConnections = mConnections;
+                // Mask-out connection direction bits to keep only Foam related connections
+                mConnections = (byte) (mMetaTileEntity.mConnections | (mConnections & ~IConnectable.CONNECTED_ALL));
+                // If foam not hardened, tries roll chance to harden
+                if ((mConnections & IConnectable.HAS_FOAM) == IConnectable.HAS_FRESHFOAM
+                    && getRandomNumber(1000) == 0) {
+                    mConnections = (byte) ((mConnections & ~IConnectable.HAS_FRESHFOAM)
+                        | IConnectable.HAS_HARDENEDFOAM);
+                }
+                if (mTickTimer > 12 && oldConnections != mConnections)
+                    GregTechAPI.causeCableUpdate(worldObj, xCoord, yCoord, zCoord);
+            }
+            mMetaTileEntity.onPreTick(this, mTickTimer);
+            if (!hasValidMetaTileEntity()) return;
+            if (isServerSide) {
+                if (mTickTimer == 10) {
+                    issueBlockUpdate();
+                    joinEnet();
+                }
+
+                if (xCoord != oX || yCoord != oY || zCoord != oZ) {
                     oX = xCoord;
                     oY = yCoord;
                     oZ = zCoord;
-                    if (isServerSide) checkDropCover();
-                    else {
-                        requestCoverDataIfNeeded();
-                    }
-                    worldObj.markTileEntityChunkModified(xCoord, yCoord, zCoord, this);
-                    mMetaTileEntity.onFirstTick(this);
-                    if (!hasValidMetaTileEntity()) return;
-                }
-
-                if (isClientSide()) {
-                    if (mColor != oColor) {
-                        mMetaTileEntity.onColorChangeClient(oColor = mColor);
-                        issueTextureUpdate();
-                    }
-
-                    if (mNeedsUpdate) {
-                        worldObj.markBlockForUpdate(xCoord, yCoord, zCoord);
-                        mNeedsUpdate = false;
-                    }
-                }
-                if (isServerSide && mTickTimer > 10) {
-                    if (!doCoverThings()) return;
-
-                    final byte oldConnections = mConnections;
-                    // Mask-out connection direction bits to keep only Foam related connections
-                    mConnections = (byte) (mMetaTileEntity.mConnections | (mConnections & ~IConnectable.CONNECTED_ALL));
-                    // If foam not hardened, tries roll chance to harden
-                    if ((mConnections & IConnectable.HAS_FOAM) == IConnectable.HAS_FRESHFOAM
-                        && getRandomNumber(1000) == 0) {
-                        mConnections = (byte) ((mConnections & ~IConnectable.HAS_FRESHFOAM)
-                            | IConnectable.HAS_HARDENEDFOAM);
-                    }
-                    if (mTickTimer > 12 && oldConnections != mConnections)
-                        GregTechAPI.causeCableUpdate(worldObj, xCoord, yCoord, zCoord);
-                }
-                mMetaTileEntity.onPreTick(this, mTickTimer);
-                if (!hasValidMetaTileEntity()) return;
-                if (isServerSide) {
-                    if (mTickTimer == 10) {
-                        issueBlockUpdate();
-                        joinEnet();
-                    }
-
-                    if (xCoord != oX || yCoord != oY || zCoord != oZ) {
-                        oX = xCoord;
-                        oY = yCoord;
-                        oZ = zCoord;
-                        issueClientUpdate();
-                        clearTileEntityBuffer();
-                    }
-                }
-
-                mMetaTileEntity.onPostTick(this, mTickTimer);
-                if (!hasValidMetaTileEntity()) return;
-
-                if (isServerSide) {
-                    if (mTickTimer % 10 == 0) {
-                        sendClientData();
-                    }
-
-                    if (mTickTimer > 10) {
-                        if (mConnections != oTextureData) sendBlockEvent((byte) 0, oTextureData = mConnections);
-                        byte tData = mMetaTileEntity.getUpdateData();
-                        if (tData != oUpdateData) sendBlockEvent((byte) 1, oUpdateData = tData);
-                        if (mColor != oColor) sendBlockEvent((byte) 2, oColor = mColor);
-                        tData = (byte) (((mSidedRedstone[0] > 0) ? 1 : 0) | ((mSidedRedstone[1] > 0) ? 2 : 0)
-                            | ((mSidedRedstone[2] > 0) ? 4 : 0)
-                            | ((mSidedRedstone[3] > 0) ? 8 : 0)
-                            | ((mSidedRedstone[4] > 0) ? 16 : 0)
-                            | ((mSidedRedstone[5] > 0) ? 32 : 0));
-                        if (tData != oRedstoneData) sendBlockEvent((byte) 3, oRedstoneData = tData);
-                    }
-
-                    if (mNeedsBlockUpdate) {
-                        updateNeighbours(mStrongRedstone, oStrongRedstone);
-                        oStrongRedstone = mStrongRedstone;
-                        mNeedsBlockUpdate = false;
-                    }
+                    issueClientUpdate();
+                    clearTileEntityBuffer();
                 }
             }
-        } catch (Throwable e) {
-            e.printStackTrace();
-            e.printStackTrace(GTLog.err);
-        }
 
-        if (isServerSide && hasTimeStatisticsStarted && hasValidMetaTileEntity()) {
-            tTime = System.nanoTime() - tTime;
-            mTimeStatisticsIndex = (mTimeStatisticsIndex + 1) % mTimeStatistics.length;
-            mTimeStatistics[mTimeStatisticsIndex] = (int) tTime;
-            if (tTime > 0 && tTime > (GregTechAPI.MILLISECOND_THRESHOLD_UNTIL_LAG_WARNING * 1000000L)
-                && mTickTimer > 1000
-                && getMetaTileEntity().doTickProfilingMessageDuringThisTick()
-                && mLagWarningCount++ < 10)
-                GT_FML_LOGGER.warn(
-                    "WARNING: Possible Lag Source at [" + xCoord
-                        + ","
-                        + yCoord
-                        + ","
-                        + zCoord
-                        + "] in Dimension "
-                        + worldObj.provider.dimensionId
-                        + " with "
-                        + tTime
-                        + " ns caused by an instance of "
-                        + getMetaTileEntity().getClass());
+            mMetaTileEntity.onPostTick(this, mTickTimer);
+            if (!hasValidMetaTileEntity()) return;
+
+            if (isServerSide) {
+                if (mTickTimer % 10 == 0) {
+                    sendClientData();
+                }
+
+                if (mTickTimer > 10) {
+                    if (mConnections != oTextureData) sendBlockEvent((byte) 0, oTextureData = mConnections);
+                    byte tData = mMetaTileEntity.getUpdateData();
+                    if (tData != oUpdateData) sendBlockEvent((byte) 1, oUpdateData = tData);
+                    if (mColor != oColor) sendBlockEvent((byte) 2, oColor = mColor);
+                    tData = (byte) (((mSidedRedstone[0] > 0) ? 1 : 0) | ((mSidedRedstone[1] > 0) ? 2 : 0)
+                        | ((mSidedRedstone[2] > 0) ? 4 : 0)
+                        | ((mSidedRedstone[3] > 0) ? 8 : 0)
+                        | ((mSidedRedstone[4] > 0) ? 16 : 0)
+                        | ((mSidedRedstone[5] > 0) ? 32 : 0));
+                    if (tData != oRedstoneData) sendBlockEvent((byte) 3, oRedstoneData = tData);
+                }
+
+                if (mNeedsBlockUpdate) {
+                    updateNeighbours(mStrongRedstone, oStrongRedstone);
+                    oStrongRedstone = mStrongRedstone;
+                    mNeedsBlockUpdate = false;
+                }
+            }
         }
 
         mWorkUpdate = mInventoryChanged = false;
@@ -431,41 +392,7 @@ public class BaseMetaPipeEntity extends CommonBaseMetaTileEntity
                         : " "));
         }
         if (aLogLevel > 1) {
-            if (hasTimeStatisticsStarted) {
-                double tAverageTime = 0;
-                double tWorstTime = 0;
-                int amountOfZero = 0;
-                for (int tTime : mTimeStatistics) {
-                    tAverageTime += tTime;
-                    if (tTime > tWorstTime) {
-                        tWorstTime = tTime;
-                    }
-                    if (tTime == 0) {
-                        amountOfZero += 1;
-                    }
-                }
-                // tick time zero means it has not been updated yet
-                int samples = mTimeStatistics.length - amountOfZero;
-                if (samples > 0) {
-                    tList.add(
-                        "Average CPU-load of ~" + (tAverageTime / samples)
-                            + "ns since "
-                            + samples
-                            + " ticks with worst time of "
-                            + tWorstTime
-                            + "ns.");
-                }
-            } else {
-                startTimeStatistics();
-                tList.add("Just started tick time statistics.");
-            }
-            if (mLagWarningCount > 0) {
-                tList.add(
-                    "Caused " + (mLagWarningCount >= 10 ? "more than 10" : mLagWarningCount)
-                        + " Lag Spike Warnings (anything taking longer than "
-                        + GregTechAPI.MILLISECOND_THRESHOLD_UNTIL_LAG_WARNING
-                        + "ms) on the Server.");
-            }
+            addProfilingInformation(tList);
             if (mMetaTileEntity != null) {
                 tList.add(
                     "Is" + (mMetaTileEntity.isAccessAllowed(aPlayer) ? " "
@@ -1292,16 +1219,6 @@ public class BaseMetaPipeEntity extends CommonBaseMetaTileEntity
     @Override
     public int getLightOpacity() {
         return mMetaTileEntity == null ? 0 : mMetaTileEntity.getLightOpacity();
-    }
-
-    @Override
-    public int[] getTimeStatistics() {
-        return mTimeStatistics;
-    }
-
-    @Override
-    public void startTimeStatistics() {
-        hasTimeStatisticsStarted = true;
     }
 
     @Override
