+++ src/main/java/gregtech/api/metatileentity/BaseMetaTileEntity.java
@@ -99,8 +99,6 @@ public class BaseMetaTileEntity extends CommonBaseMetaTileEntity
         .findField(EntityItem.class, "health", "field_70291_e");
     private final boolean[] mActiveEUInputs = new boolean[] { false, false, false, false, false, false };
     private final boolean[] mActiveEUOutputs = new boolean[] { false, false, false, false, false, false };
-    private final int[] mTimeStatistics = new int[GregTechAPI.TICKS_FOR_LAG_AVERAGING];
-    private boolean hasTimeStatisticsStarted;
     public long mLastSoundTick = 0;
     public boolean mWasShutdown = false;
     public @Nonnull ShutDownReason lastShutDownReason = ShutDownReasonRegistry.NONE;
@@ -118,7 +116,7 @@ public class BaseMetaTileEntity extends CommonBaseMetaTileEntity
         oTexturePage = 0;
     private byte oLightValueClient = 0, oLightValue = -1, mLightValue = 0, mOtherUpgrades = 0;
     private ForgeDirection mFacing = ForgeDirection.DOWN, oFacing = ForgeDirection.DOWN;
-    private int oX = 0, oY = 0, oZ = 0, mTimeStatisticsIndex = 0, mLagWarningCount = 0;
+    private int oX = 0, oY = 0, oZ = 0;
     private long oOutput = 0, mAcceptedAmperes = Long.MAX_VALUE;
     private long mLastCheckTick = 0;
     private String mOwnerName = "";
@@ -284,201 +282,154 @@ public class BaseMetaTileEntity extends CommonBaseMetaTileEntity
     }
 
     @Override
-    public void updateEntity() {
-        super.updateEntity();
-
+    public void updateEntityProfiled() {
         if (!hasValidMetaTileEntity()) {
             if (mMetaTileEntity == null) return;
             mMetaTileEntity.setBaseMetaTileEntity(this);
         }
 
         mRunningThroughTick = true;
-        long tTime;
-        if (hasTimeStatisticsStarted) {
-            tTime = System.nanoTime();
-        } else {
-            tTime = 0;
-        }
         final boolean isServerSide = isServerSide();
-
-        try {
-            if (hasValidMetaTileEntity()) {
-                if (mTickTimer++ == 0) {
-                    oX = xCoord;
-                    oY = yCoord;
-                    oZ = zCoord;
-                    if (isServerSide) {
-                        checkDropCover();
-                    } else {
-                        requestCoverDataIfNeeded();
-                    }
-                    worldObj.markTileEntityChunkModified(xCoord, yCoord, zCoord, this);
-                    mMetaTileEntity.onFirstTick(this);
-                    if (!hasValidMetaTileEntity()) {
-                        mRunningThroughTick = false;
-                        return;
-                    }
-                }
-                if (isClientSide()) {
-                    if (mColor != oColor) {
-                        mMetaTileEntity.onColorChangeClient(oColor = mColor);
-                        issueTextureUpdate();
-                    }
-
-                    if (mLightValue != oLightValueClient) {
-                        worldObj.setLightValue(EnumSkyBlock.Block, xCoord, yCoord, zCoord, mLightValue);
-                        worldObj.updateLightByType(EnumSkyBlock.Block, xCoord, yCoord, zCoord);
-                        worldObj.updateLightByType(EnumSkyBlock.Block, xCoord + 1, yCoord, zCoord);
-                        worldObj.updateLightByType(EnumSkyBlock.Block, xCoord - 1, yCoord, zCoord);
-                        worldObj.updateLightByType(EnumSkyBlock.Block, xCoord, yCoord + 1, zCoord);
-                        worldObj.updateLightByType(EnumSkyBlock.Block, xCoord, yCoord - 1, zCoord);
-                        worldObj.updateLightByType(EnumSkyBlock.Block, xCoord, yCoord, zCoord + 1);
-                        worldObj.updateLightByType(EnumSkyBlock.Block, xCoord, yCoord, zCoord - 1);
-                        oLightValueClient = mLightValue;
-                        issueTextureUpdate();
-                    }
-
-                    if (mNeedsUpdate) {
-                        if (GTMod.gregtechproxy.mUseBlockUpdateHandler) {
-                            BlockUpdateHandler.Instance.enqueueBlockUpdate(worldObj, getLocation());
-                        } else {
-                            worldObj.markBlockForUpdate(xCoord, yCoord, zCoord);
-                        }
-                        mNeedsUpdate = false;
-                    }
-                }
-                if (isServerSide && mTickTimer > 10) {
-                    if (!doCoverThings()) {
-                        mRunningThroughTick = false;
-                        return;
-                    }
-                }
+        if (hasValidMetaTileEntity()) {
+            if (mTickTimer++ == 0) {
+                oX = xCoord;
+                oY = yCoord;
+                oZ = zCoord;
                 if (isServerSide) {
-                    if (++mAverageEUInputIndex >= mAverageEUInput.length) mAverageEUInputIndex = 0;
-                    if (++mAverageEUOutputIndex >= mAverageEUOutput.length) mAverageEUOutputIndex = 0;
-
-                    mAverageEUInput[mAverageEUInputIndex] = 0;
-                    mAverageEUOutput[mAverageEUOutputIndex] = 0;
+                    checkDropCover();
+                } else {
+                    requestCoverDataIfNeeded();
                 }
-
-                mMetaTileEntity.onPreTick(this, mTickTimer);
-
+                worldObj.markTileEntityChunkModified(xCoord, yCoord, zCoord, this);
+                mMetaTileEntity.onFirstTick(this);
                 if (!hasValidMetaTileEntity()) {
                     mRunningThroughTick = false;
                     return;
                 }
-                if (isServerSide) {
-                    if (mRedstone != oRedstone || mTickTimer == 10) {
-                        oRedstone = mRedstone;
-                        issueBlockUpdate();
+            }
+            if (isClientSide()) {
+                if (mColor != oColor) {
+                    mMetaTileEntity.onColorChangeClient(oColor = mColor);
+                    issueTextureUpdate();
+                }
+
+                if (mLightValue != oLightValueClient) {
+                    worldObj.setLightValue(EnumSkyBlock.Block, xCoord, yCoord, zCoord, mLightValue);
+                    worldObj.updateLightByType(EnumSkyBlock.Block, xCoord, yCoord, zCoord);
+                    worldObj.updateLightByType(EnumSkyBlock.Block, xCoord + 1, yCoord, zCoord);
+                    worldObj.updateLightByType(EnumSkyBlock.Block, xCoord - 1, yCoord, zCoord);
+                    worldObj.updateLightByType(EnumSkyBlock.Block, xCoord, yCoord + 1, zCoord);
+                    worldObj.updateLightByType(EnumSkyBlock.Block, xCoord, yCoord - 1, zCoord);
+                    worldObj.updateLightByType(EnumSkyBlock.Block, xCoord, yCoord, zCoord + 1);
+                    worldObj.updateLightByType(EnumSkyBlock.Block, xCoord, yCoord, zCoord - 1);
+                    oLightValueClient = mLightValue;
+                    issueTextureUpdate();
+                }
+
+                if (mNeedsUpdate) {
+                    if (GTMod.gregtechproxy.mUseBlockUpdateHandler) {
+                        BlockUpdateHandler.Instance.enqueueBlockUpdate(worldObj, getLocation());
+                    } else {
+                        worldObj.markBlockForUpdate(xCoord, yCoord, zCoord);
                     }
-                    if (mTickTimer == 10) joinEnet();
+                    mNeedsUpdate = false;
+                }
+            }
+            if (isServerSide && mTickTimer > 10) {
+                if (!doCoverThings()) {
+                    mRunningThroughTick = false;
+                    return;
+                }
+            }
+            if (isServerSide) {
+                if (++mAverageEUInputIndex >= mAverageEUInput.length) mAverageEUInputIndex = 0;
+                if (++mAverageEUOutputIndex >= mAverageEUOutput.length) mAverageEUOutputIndex = 0;
 
-                    if (xCoord != oX || yCoord != oY || zCoord != oZ) {
-                        oX = xCoord;
-                        oY = yCoord;
-                        oZ = zCoord;
-                        issueClientUpdate();
-                        clearTileEntityBuffer();
+                mAverageEUInput[mAverageEUInputIndex] = 0;
+                mAverageEUOutput[mAverageEUOutputIndex] = 0;
+            }
+
+            mMetaTileEntity.onPreTick(this, mTickTimer);
+
+            if (!hasValidMetaTileEntity()) {
+                mRunningThroughTick = false;
+                return;
+            }
+            if (isServerSide) {
+                if (mRedstone != oRedstone || mTickTimer == 10) {
+                    oRedstone = mRedstone;
+                    issueBlockUpdate();
+                }
+                if (mTickTimer == 10) joinEnet();
+
+                if (xCoord != oX || yCoord != oY || zCoord != oZ) {
+                    oX = xCoord;
+                    oY = yCoord;
+                    oZ = zCoord;
+                    issueClientUpdate();
+                    clearTileEntityBuffer();
+                }
+
+                if (mFacing != oFacing) {
+                    oFacing = mFacing;
+                    checkDropCover();
+                    issueBlockUpdate();
+                }
+
+                if (mTickTimer > 20 && mMetaTileEntity.isElectric()) {
+                    mAcceptedAmperes = 0;
+
+                    if (getOutputVoltage() != oOutput) {
+                        oOutput = getOutputVoltage();
                     }
 
-                    if (mFacing != oFacing) {
-                        oFacing = mFacing;
-                        checkDropCover();
-                        issueBlockUpdate();
+                    if (mMetaTileEntity.isEnetOutput() || mMetaTileEntity.isEnetInput()) {
+                        for (final ForgeDirection side : ForgeDirection.VALID_DIRECTIONS) {
+                            final int ordinalSide = side.ordinal();
+                            boolean temp = isEnergyInputSide(side);
+                            if (temp != mActiveEUInputs[ordinalSide]) {
+                                mActiveEUInputs[ordinalSide] = temp;
+                            }
+                            temp = isEnergyOutputSide(side);
+                            if (temp != mActiveEUOutputs[ordinalSide]) {
+                                mActiveEUOutputs[ordinalSide] = temp;
+                            }
+                        }
                     }
 
-                    if (mTickTimer > 20 && mMetaTileEntity.isElectric()) {
-                        mAcceptedAmperes = 0;
-
-                        if (getOutputVoltage() != oOutput) {
-                            oOutput = getOutputVoltage();
+                    if (mMetaTileEntity.isEnetOutput() && oOutput > 0) {
+                        final long tOutputVoltage = Math
+                            .max(oOutput, oOutput + (1L << Math.max(0, GTUtility.getTier(oOutput) - 1)));
+                        final long tUsableAmperage = Math.min(
+                            getOutputAmperage(),
+                            (getStoredEU() - mMetaTileEntity.getMinimumStoredEU()) / tOutputVoltage);
+                        if (tUsableAmperage > 0) {
+                            final long tEU = tOutputVoltage * Util.emitEnergyToNetwork(oOutput, tUsableAmperage, this);
+                            mAverageEUOutput[mAverageEUOutputIndex] += tEU;
+                            decreaseStoredEU(tEU, true);
+                        }
+                    }
+                    if (getEUCapacity() > 0) {
+                        if (GregTechAPI.sMachineFireExplosions && getRandomNumber(1000) == 0) {
+                            final Block tBlock = getBlockAtSide(ForgeDirection.getOrientation(getRandomNumber(6)));
+                            if (tBlock instanceof BlockFire) doEnergyExplosion();
                         }
 
-                        if (mMetaTileEntity.isEnetOutput() || mMetaTileEntity.isEnetInput()) {
-                            for (final ForgeDirection side : ForgeDirection.VALID_DIRECTIONS) {
-                                final int ordinalSide = side.ordinal();
-                                boolean temp = isEnergyInputSide(side);
-                                if (temp != mActiveEUInputs[ordinalSide]) {
-                                    mActiveEUInputs[ordinalSide] = temp;
-                                }
-                                temp = isEnergyOutputSide(side);
-                                if (temp != mActiveEUOutputs[ordinalSide]) {
-                                    mActiveEUOutputs[ordinalSide] = temp;
-                                }
-                            }
+                        if (!hasValidMetaTileEntity()) {
+                            mRunningThroughTick = false;
+                            return;
                         }
 
-                        if (mMetaTileEntity.isEnetOutput() && oOutput > 0) {
-                            final long tOutputVoltage = Math
-                                .max(oOutput, oOutput + (1L << Math.max(0, GTUtility.getTier(oOutput) - 1)));
-                            final long tUsableAmperage = Math.min(
-                                getOutputAmperage(),
-                                (getStoredEU() - mMetaTileEntity.getMinimumStoredEU()) / tOutputVoltage);
-                            if (tUsableAmperage > 0) {
-                                final long tEU = tOutputVoltage
-                                    * Util.emitEnergyToNetwork(oOutput, tUsableAmperage, this);
-                                mAverageEUOutput[mAverageEUOutputIndex] += tEU;
-                                decreaseStoredEU(tEU, true);
-                            }
-                        }
-                        if (getEUCapacity() > 0) {
-                            if (GregTechAPI.sMachineFireExplosions && getRandomNumber(1000) == 0) {
-                                final Block tBlock = getBlockAtSide(ForgeDirection.getOrientation(getRandomNumber(6)));
-                                if (tBlock instanceof BlockFire) doEnergyExplosion();
-                            }
-
-                            if (!hasValidMetaTileEntity()) {
-                                mRunningThroughTick = false;
-                                return;
-                            }
-
-                            if (GregTechAPI.sMachineRainExplosions) {
-                                if (mMetaTileEntity.willExplodeInRain()) {
-                                    if (getRandomNumber(1000) == 0 && isRainPossible()) {
-                                        // Short-circuit so raincheck happens before isRainExposed,
-                                        // saves sme TPS since rain exposed check can be slow
-                                        // This logic can be compressed further by only checking for
-                                        // isRainExposed once IF we can guarantee it never thunders without
-                                        // raining, but I don't know if this is true or not.
-                                        if (worldObj.isRaining() && isRainExposed()) {
-                                            if (getRandomNumber(10) == 0) {
-                                                try {
-                                                    GTMod.achievements.issueAchievement(
-                                                        this.getWorldObj()
-                                                            .getPlayerEntityByName(mOwnerName),
-                                                        "badweather");
-                                                } catch (Exception ignored) {}
-                                                GTLog.exp.println(
-                                                    "Machine at: " + this.getXCoord()
-                                                        + " | "
-                                                        + this.getYCoord()
-                                                        + " | "
-                                                        + this.getZCoord()
-                                                        + " DIMID: "
-                                                        + this.worldObj.provider.dimensionId
-                                                        + " explosion due to rain!");
-                                                doEnergyExplosion();
-                                            } else {
-                                                GTLog.exp.println(
-                                                    "Machine at: " + this.getXCoord()
-                                                        + " | "
-                                                        + this.getYCoord()
-                                                        + " | "
-                                                        + this.getZCoord()
-                                                        + " DIMID: "
-                                                        + this.worldObj.provider.dimensionId
-                                                        + "  set to Fire due to rain!");
-                                                setOnFire();
-                                            }
-                                        }
-                                        if (!hasValidMetaTileEntity()) {
-                                            mRunningThroughTick = false;
-                                            return;
-                                        }
-                                        if (GregTechAPI.sMachineThunderExplosions && worldObj.isThundering()
-                                            && getRandomNumber(3) == 0
-                                            && isRainExposed()) {
+                        if (GregTechAPI.sMachineRainExplosions) {
+                            if (mMetaTileEntity.willExplodeInRain()) {
+                                if (getRandomNumber(1000) == 0 && isRainPossible()) {
+                                    // Short-circuit so raincheck happens before isRainExposed,
+                                    // saves sme TPS since rain exposed check can be slow
+                                    // This logic can be compressed further by only checking for
+                                    // isRainExposed once IF we can guarantee it never thunders without
+                                    // raining, but I don't know if this is true or not.
+                                    if (worldObj.isRaining() && isRainExposed()) {
+                                        if (getRandomNumber(10) == 0) {
                                             try {
                                                 GTMod.achievements.issueAchievement(
                                                     this.getWorldObj()
@@ -493,157 +444,161 @@ public class BaseMetaTileEntity extends CommonBaseMetaTileEntity
                                                     + this.getZCoord()
                                                     + " DIMID: "
                                                     + this.worldObj.provider.dimensionId
-                                                    + " explosion due to Thunderstorm!");
+                                                    + " explosion due to rain!");
                                             doEnergyExplosion();
+                                        } else {
+                                            GTLog.exp.println(
+                                                "Machine at: " + this.getXCoord()
+                                                    + " | "
+                                                    + this.getYCoord()
+                                                    + " | "
+                                                    + this.getZCoord()
+                                                    + " DIMID: "
+                                                    + this.worldObj.provider.dimensionId
+                                                    + "  set to Fire due to rain!");
+                                            setOnFire();
                                         }
                                     }
-                                }
-                            }
-                        }
-                    }
-
-                    if (!hasValidMetaTileEntity()) {
-                        mRunningThroughTick = false;
-                        return;
-                    }
-                }
-                if (isServerSide) {
-                    if (mMetaTileEntity.dechargerSlotCount() > 0 && getStoredEU() < getEUCapacity()) {
-                        for (int i = mMetaTileEntity.dechargerSlotStartIndex(),
-                            k = mMetaTileEntity.dechargerSlotCount() + i; i < k; i++) {
-                            if (mMetaTileEntity.mInventory[i] != null && getStoredEU() < getEUCapacity()) {
-                                dischargeItem(mMetaTileEntity.mInventory[i]);
-                                if (ic2.api.info.Info.itemEnergy.getEnergyValue(mMetaTileEntity.mInventory[i]) > 0) {
-                                    if ((getStoredEU()
-                                        + ic2.api.info.Info.itemEnergy.getEnergyValue(mMetaTileEntity.mInventory[i]))
-                                        < getEUCapacity()) {
-                                        increaseStoredEnergyUnits(
-                                            (long) ic2.api.info.Info.itemEnergy
-                                                .getEnergyValue(mMetaTileEntity.mInventory[i]),
-                                            false);
-                                        mMetaTileEntity.mInventory[i].stackSize--;
-                                        mInventoryChanged = true;
+                                    if (!hasValidMetaTileEntity()) {
+                                        mRunningThroughTick = false;
+                                        return;
+                                    }
+                                    if (GregTechAPI.sMachineThunderExplosions && worldObj.isThundering()
+                                        && getRandomNumber(3) == 0
+                                        && isRainExposed()) {
+                                        try {
+                                            GTMod.achievements.issueAchievement(
+                                                this.getWorldObj()
+                                                    .getPlayerEntityByName(mOwnerName),
+                                                "badweather");
+                                        } catch (Exception ignored) {}
+                                        GTLog.exp.println(
+                                            "Machine at: " + this.getXCoord()
+                                                + " | "
+                                                + this.getYCoord()
+                                                + " | "
+                                                + this.getZCoord()
+                                                + " DIMID: "
+                                                + this.worldObj.provider.dimensionId
+                                                + " explosion due to Thunderstorm!");
+                                        doEnergyExplosion();
                                     }
                                 }
-                                if (mMetaTileEntity.mInventory[i].stackSize <= 0) {
-                                    mMetaTileEntity.mInventory[i] = null;
-                                    mInventoryChanged = true;
-                                }
                             }
                         }
                     }
                 }
-                if (isServerSide) {
-                    if (mMetaTileEntity.rechargerSlotCount() > 0 && getStoredEU() > 0) {
-                        for (int i = mMetaTileEntity.rechargerSlotStartIndex(),
-                            k = mMetaTileEntity.rechargerSlotCount() + i; i < k; i++) {
-                            if (getStoredEU() > 0 && mMetaTileEntity.mInventory[i] != null) {
-                                chargeItem(mMetaTileEntity.mInventory[i]);
-                                if (mMetaTileEntity.mInventory[i].stackSize <= 0) {
-                                    mMetaTileEntity.mInventory[i] = null;
+
+                if (!hasValidMetaTileEntity()) {
+                    mRunningThroughTick = false;
+                    return;
+                }
+            }
+            if (isServerSide) {
+                if (mMetaTileEntity.dechargerSlotCount() > 0 && getStoredEU() < getEUCapacity()) {
+                    for (int i = mMetaTileEntity.dechargerSlotStartIndex(),
+                        k = mMetaTileEntity.dechargerSlotCount() + i; i < k; i++) {
+                        if (mMetaTileEntity.mInventory[i] != null && getStoredEU() < getEUCapacity()) {
+                            dischargeItem(mMetaTileEntity.mInventory[i]);
+                            if (ic2.api.info.Info.itemEnergy.getEnergyValue(mMetaTileEntity.mInventory[i]) > 0) {
+                                if ((getStoredEU()
+                                    + ic2.api.info.Info.itemEnergy.getEnergyValue(mMetaTileEntity.mInventory[i]))
+                                    < getEUCapacity()) {
+                                    increaseStoredEnergyUnits(
+                                        (long) ic2.api.info.Info.itemEnergy
+                                            .getEnergyValue(mMetaTileEntity.mInventory[i]),
+                                        false);
+                                    mMetaTileEntity.mInventory[i].stackSize--;
                                     mInventoryChanged = true;
                                 }
+                            }
+                            if (mMetaTileEntity.mInventory[i].stackSize <= 0) {
+                                mMetaTileEntity.mInventory[i] = null;
+                                mInventoryChanged = true;
                             }
                         }
                     }
                 }
-                updateStatus();
-                if (!hasValidMetaTileEntity()) {
-                    mRunningThroughTick = false;
-                    return;
-                }
-                mMetaTileEntity.onPostTick(this, mTickTimer);
-                if (!hasValidMetaTileEntity()) {
-                    mRunningThroughTick = false;
-                    return;
-                }
-                if (isServerSide) {
-                    if (mTickTimer > 20 && cableUpdateDelay == 0) {
-                        generatePowerNodes();
-                    }
-                    cableUpdateDelay--;
-                    if (mTickTimer % 10 == 0) {
-                        sendClientData();
-                    }
-
-                    if (mTickTimer > 10) {
-                        byte tData = (byte) ((mFacing.ordinal() & 7) | (mActive ? 8 : 0)
-                            | (mRedstone ? 16 : 0)
-                            | (mLockUpgrade ? 32 : 0)
-                            | (mWorks ? 64 : 0)
-                            | (mMuffler ? 128 : 0));
-                        if (tData != oTextureData)
-                            sendBlockEvent(GregTechTileClientEvents.CHANGE_COMMON_DATA, oTextureData = tData);
-
-                        tData = mMetaTileEntity.getUpdateData();
-                        if (tData != oUpdateData)
-                            sendBlockEvent(GregTechTileClientEvents.CHANGE_CUSTOM_DATA, oUpdateData = tData);
-                        if (mMetaTileEntity instanceof MTEHatch) {
-                            tData = ((MTEHatch) mMetaTileEntity).getTexturePage();
-                            if (tData != oTexturePage) sendBlockEvent(
-                                GregTechTileClientEvents.CHANGE_CUSTOM_DATA,
-                                (byte) ((oTexturePage = tData) | 0x80)); // set last bit as a flag for page
+            }
+            if (isServerSide) {
+                if (mMetaTileEntity.rechargerSlotCount() > 0 && getStoredEU() > 0) {
+                    for (int i = mMetaTileEntity.rechargerSlotStartIndex(),
+                        k = mMetaTileEntity.rechargerSlotCount() + i; i < k; i++) {
+                        if (getStoredEU() > 0 && mMetaTileEntity.mInventory[i] != null) {
+                            chargeItem(mMetaTileEntity.mInventory[i]);
+                            if (mMetaTileEntity.mInventory[i].stackSize <= 0) {
+                                mMetaTileEntity.mInventory[i] = null;
+                                mInventoryChanged = true;
+                            }
                         }
-                        if (mColor != oColor) sendBlockEvent(GregTechTileClientEvents.CHANGE_COLOR, oColor = mColor);
-                        tData = (byte) (((mSidedRedstone[0] > 0) ? 1 : 0) | ((mSidedRedstone[1] > 0) ? 2 : 0)
-                            | ((mSidedRedstone[2] > 0) ? 4 : 0)
-                            | ((mSidedRedstone[3] > 0) ? 8 : 0)
-                            | ((mSidedRedstone[4] > 0) ? 16 : 0)
-                            | ((mSidedRedstone[5] > 0) ? 32 : 0));
-                        if (tData != oRedstoneData)
-                            sendBlockEvent(GregTechTileClientEvents.CHANGE_REDSTONE_OUTPUT, oRedstoneData = tData);
-                        if (mLightValue != oLightValue) {
-                            worldObj.setLightValue(EnumSkyBlock.Block, xCoord, yCoord, zCoord, mLightValue);
-                            worldObj.updateLightByType(EnumSkyBlock.Block, xCoord, yCoord, zCoord);
-                            worldObj.updateLightByType(EnumSkyBlock.Block, xCoord + 1, yCoord, zCoord);
-                            worldObj.updateLightByType(EnumSkyBlock.Block, xCoord - 1, yCoord, zCoord);
-                            worldObj.updateLightByType(EnumSkyBlock.Block, xCoord, yCoord + 1, zCoord);
-                            worldObj.updateLightByType(EnumSkyBlock.Block, xCoord, yCoord - 1, zCoord);
-                            worldObj.updateLightByType(EnumSkyBlock.Block, xCoord, yCoord, zCoord + 1);
-                            worldObj.updateLightByType(EnumSkyBlock.Block, xCoord, yCoord, zCoord - 1);
-                            issueTextureUpdate();
-                            sendBlockEvent(GregTechTileClientEvents.CHANGE_LIGHT, oLightValue = mLightValue);
-                        }
-                    }
-
-                    if (mNeedsBlockUpdate) {
-                        updateNeighbours(mStrongRedstone, oStrongRedstone);
-                        oStrongRedstone = mStrongRedstone;
-                        mNeedsBlockUpdate = false;
                     }
                 }
             }
-        } catch (Throwable e) {
-            e.printStackTrace();
-            e.printStackTrace(GTLog.err);
-            try {
-                mMetaTileEntity.onTickFail(this, mTickTimer);
-            } catch (Throwable ex) {
-                ex.printStackTrace();
-                ex.printStackTrace(GTLog.err);
+            updateStatus();
+            if (!hasValidMetaTileEntity()) {
+                mRunningThroughTick = false;
+                return;
             }
-        }
+            mMetaTileEntity.onPostTick(this, mTickTimer);
+            if (!hasValidMetaTileEntity()) {
+                mRunningThroughTick = false;
+                return;
+            }
+            if (isServerSide) {
+                if (mTickTimer > 20 && cableUpdateDelay == 0) {
+                    generatePowerNodes();
+                }
+                cableUpdateDelay--;
+                if (mTickTimer % 10 == 0) {
+                    sendClientData();
+                }
 
-        if (isServerSide && hasTimeStatisticsStarted && hasValidMetaTileEntity()) {
-            tTime = System.nanoTime() - tTime;
-            mTimeStatisticsIndex = (mTimeStatisticsIndex + 1) % mTimeStatistics.length;
-            mTimeStatistics[mTimeStatisticsIndex] = (int) tTime;
-            if (tTime > 0 && tTime > (GregTechAPI.MILLISECOND_THRESHOLD_UNTIL_LAG_WARNING * 1_000_000L)
-                && mTickTimer > 1000
-                && getMetaTileEntity().doTickProfilingMessageDuringThisTick()
-                && mLagWarningCount++ < 10)
-                GT_FML_LOGGER.warn(
-                    "WARNING: Possible Lag Source at [" + xCoord
-                        + ", "
-                        + yCoord
-                        + ", "
-                        + zCoord
-                        + "] in Dimension "
-                        + worldObj.provider.dimensionId
-                        + " with "
-                        + tTime
-                        + "ns caused by an instance of "
-                        + getMetaTileEntity().getClass());
+                if (mTickTimer > 10) {
+                    byte tData = (byte) ((mFacing.ordinal() & 7) | (mActive ? 8 : 0)
+                        | (mRedstone ? 16 : 0)
+                        | (mLockUpgrade ? 32 : 0)
+                        | (mWorks ? 64 : 0)
+                        | (mMuffler ? 128 : 0));
+                    if (tData != oTextureData)
+                        sendBlockEvent(GregTechTileClientEvents.CHANGE_COMMON_DATA, oTextureData = tData);
+
+                    tData = mMetaTileEntity.getUpdateData();
+                    if (tData != oUpdateData)
+                        sendBlockEvent(GregTechTileClientEvents.CHANGE_CUSTOM_DATA, oUpdateData = tData);
+                    if (mMetaTileEntity instanceof MTEHatch) {
+                        tData = ((MTEHatch) mMetaTileEntity).getTexturePage();
+                        if (tData != oTexturePage) sendBlockEvent(
+                            GregTechTileClientEvents.CHANGE_CUSTOM_DATA,
+                            (byte) ((oTexturePage = tData) | 0x80)); // set last bit as a flag for page
+                    }
+                    if (mColor != oColor) sendBlockEvent(GregTechTileClientEvents.CHANGE_COLOR, oColor = mColor);
+                    tData = (byte) (((mSidedRedstone[0] > 0) ? 1 : 0) | ((mSidedRedstone[1] > 0) ? 2 : 0)
+                        | ((mSidedRedstone[2] > 0) ? 4 : 0)
+                        | ((mSidedRedstone[3] > 0) ? 8 : 0)
+                        | ((mSidedRedstone[4] > 0) ? 16 : 0)
+                        | ((mSidedRedstone[5] > 0) ? 32 : 0));
+                    if (tData != oRedstoneData)
+                        sendBlockEvent(GregTechTileClientEvents.CHANGE_REDSTONE_OUTPUT, oRedstoneData = tData);
+                    if (mLightValue != oLightValue) {
+                        worldObj.setLightValue(EnumSkyBlock.Block, xCoord, yCoord, zCoord, mLightValue);
+                        worldObj.updateLightByType(EnumSkyBlock.Block, xCoord, yCoord, zCoord);
+                        worldObj.updateLightByType(EnumSkyBlock.Block, xCoord + 1, yCoord, zCoord);
+                        worldObj.updateLightByType(EnumSkyBlock.Block, xCoord - 1, yCoord, zCoord);
+                        worldObj.updateLightByType(EnumSkyBlock.Block, xCoord, yCoord + 1, zCoord);
+                        worldObj.updateLightByType(EnumSkyBlock.Block, xCoord, yCoord - 1, zCoord);
+                        worldObj.updateLightByType(EnumSkyBlock.Block, xCoord, yCoord, zCoord + 1);
+                        worldObj.updateLightByType(EnumSkyBlock.Block, xCoord, yCoord, zCoord - 1);
+                        issueTextureUpdate();
+                        sendBlockEvent(GregTechTileClientEvents.CHANGE_LIGHT, oLightValue = mLightValue);
+                    }
+                }
+
+                if (mNeedsBlockUpdate) {
+                    updateNeighbours(mStrongRedstone, oStrongRedstone);
+                    oStrongRedstone = mStrongRedstone;
+                    mNeedsBlockUpdate = false;
+                }
+            }
         }
 
         mWorkUpdate = mInventoryChanged = mRunningThroughTick = false;
@@ -799,36 +754,10 @@ public class BaseMetaTileEntity extends CommonBaseMetaTileEntity
                         : " "));
         }
         if (aLogLevel > 1 && mMetaTileEntity != null) {
-            if (hasTimeStatisticsStarted) {
-                double tAverageTime = 0;
-                double tWorstTime = 0;
-                int amountOfZero = 0;
-                for (int tTime : mTimeStatistics) {
-                    tAverageTime += tTime;
-                    if (tTime > tWorstTime) {
-                        tWorstTime = tTime;
-                    }
-                    if (tTime == 0) {
-                        amountOfZero += 1;
-                    }
-                    // Uncomment this line to print out tick-by-tick times.
-                    // tList.add("tTime " + tTime);
-                }
-                // tick time zero means it has not been updated yet
-                int samples = mTimeStatistics.length - amountOfZero;
-                if (samples > 0) {
-                    tList.add(
-                        "Average CPU load of ~" + GTUtility.formatNumbers(tAverageTime / samples)
-                            + "ns over "
-                            + GTUtility.formatNumbers(samples)
-                            + " ticks with worst time of "
-                            + GTUtility.formatNumbers(tWorstTime)
-                            + "ns.");
-                }
-            } else {
-                startTimeStatistics();
-                tList.add("Just started tick time statistics.");
-            }
+            addProfilingInformation(tList);
+            tList.add(
+                "Is" + (mMetaTileEntity.isAccessAllowed(aPlayer) ? " "
+                    : EnumChatFormatting.RED + " not " + EnumChatFormatting.RESET) + "accessible for you");
             tList.add(
                 "Recorded " + GTUtility.formatNumbers(mMetaTileEntity.mSoundRequests)
                     + " sound requests in "
@@ -836,16 +765,6 @@ public class BaseMetaTileEntity extends CommonBaseMetaTileEntity
                     + " ticks.");
             mLastCheckTick = mTickTimer;
             mMetaTileEntity.mSoundRequests = 0;
-            if (mLagWarningCount > 0) {
-                tList.add(
-                    "Caused " + (mLagWarningCount >= 10 ? "more than 10" : mLagWarningCount)
-                        + " Lag Spike Warnings (anything taking longer than "
-                        + GregTechAPI.MILLISECOND_THRESHOLD_UNTIL_LAG_WARNING
-                        + "ms) on the Server.");
-            }
-            tList.add(
-                "Is" + (mMetaTileEntity.isAccessAllowed(aPlayer) ? " "
-                    : EnumChatFormatting.RED + " not " + EnumChatFormatting.RESET) + "accessible for you");
         }
         if (aLogLevel > 0) {
             if (getSteamCapacity() > 0 && hasSteamEngineUpgrade()) tList.add(
@@ -2302,16 +2221,6 @@ public class BaseMetaTileEntity extends CommonBaseMetaTileEntity
     @Override
     public IConstructable getConstructable() {
         return getMetaTileEntity() instanceof IConstructable ? (IConstructable) getMetaTileEntity() : null;
-    }
-
-    @Override
-    public int[] getTimeStatistics() {
-        return mTimeStatistics;
-    }
-
-    @Override
-    public void startTimeStatistics() {
-        hasTimeStatisticsStarted = true;
     }
 
     @Nullable
